# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Glitch API",
        "description": "The API for connecting to Glitch",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/competitions/{uuid}/invites": {
            "get": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Display a list of invites.",
                "description": "Display a list of invites.",
                "operationId": "competitionUserInviteList",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompetitionInvite"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "The competition does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accessed denied to listing invites.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Accessed denied to listing invites."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/competitions/{uuid}/sendInvite": {
            "post": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Send an invite to a user.",
                "description": "Send an invite to a user.",
                "operationId": "competitionSendInvite",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionInvite"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionInvite"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "The competition does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No invites listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Cannot send invite to user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/competitions/{uuid}/acceptInvite": {
            "post": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Use this endpoint to accept the invite. Requires the users auth token to validate.",
                "description": "Use this endpoint to accept the invite. Requires the users auth token to validate.",
                "operationId": "competitionAcceptInvite",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "The token that was created with the invite.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionInvite"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "The competition does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No invites listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Invite is not associated with the current user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "No invites listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Invite does not belong to this competition."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompetitionInvite": {
                "title": "Competition Invite Model",
                "description": "Keep track of users when that have been invited to a competition.",
                "required": [
                    "name",
                    "email",
                    "competition_id"
                ],
                "properties": {
                    "id": {
                        "description": "The uuid of invite.",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "email": {
                        "description": "The email address of the person that is invited.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "name": {
                        "description": "The name of the person that is invited.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "competition_id": {
                        "description": "The id of the competition that the invite belongs too.",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "invited_as_participant": {
                        "description": "Sets if the user is to be invited as an individual participant.",
                        "type": "boolean"
                    },
                    "invited_as_team_member": {
                        "description": "Sets if the user is to be invited as part of a team.",
                        "type": "boolean"
                    },
                    "accepted_invite": {
                        "description": "Is true of the user has accepted the invite.",
                        "type": "boolean"
                    },
                    "token": {
                        "description": "The auto-generated token that is sent along with the invite.",
                        "type": "string",
                        "maxLength": 255,
                        "readOnly": true
                    },
                    "role": {
                        "description": "The role the user is being invited as.",
                        "type": "integer",
                        "enum": [
                            "0",
                            "1",
                            "2"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    }
}